Extract `load_region` into separate module (ref #616)
