+++ fractal_tasks_core/lib_regions_of_interest/__init__.py
@@ -1,6 +1,7 @@
 """
 Subpackage for ROI-related functions.
 """
+from .load_region import load_region  # noqa
 from .v1 import *  # noqa
 from .v1_checks import *  # noqa
 from .v1_overlaps import *  # noqa
